code refactored and changed points for tasks to 5
