+++ apps/courses/models.py
@@ -81,7 +81,6 @@ class Task(AbstractStatusField):
     name = models.CharField(max_length=50, verbose_name="Название")
     skill = models.ForeignKey(Skill, on_delete=models.CASCADE, related_name="tasks", verbose_name="Навык")
     level = models.IntegerField(default=1, verbose_name="Уровень")
-    # TODO добавить порядковый номер для показа
 
     def __str__(self):
         return f"name:<{self.name}> skill:<{self.skill.name}> level:<{self.level}>"
@@ -91,6 +90,7 @@ class Task(AbstractStatusField):
         verbose_name_plural = "Задачи"
 
     def save(self, *args, **kwargs):
+        """Автоматически устанавливает порядковый номер"""
         if self.ordinal_number is None:  # если порядковый номер слайда не введен
             last_task_obj = Task.objects.aggregate(Max("ordinal_number"))
             if not last_task_obj.get("ordinal_number__max", 0):
@@ -164,8 +164,9 @@ class Popup(models.Model):
 
     def clean(self):
         if not self.title and not self.text and not self.file:
-            raise ValidationError("Должено быть заполнено хотя бы один из полей: "
-                                  "'Заголовок', 'Содержимое' или 'Изображение'")
+            raise ValidationError(
+                "Должено быть заполнено хотя бы один из полей: " "'Заголовок', 'Содержимое' или 'Изображение'"
+            )
 
     def save(self, *args, **kwargs):
         # TODO автоинкремен и валидация при добалении в TaskObject
