--- apps/progress/manager.py
@@ -1,31 +1,0 @@
-from django.contrib.auth.hashers import make_password
-from django.contrib.auth.models import UserManager
-
-
-class CustomUserManager(UserManager):
-    def create_user(self, email, password=None, **extra_fields):
-        extra_fields.setdefault("is_staff", False)
-        extra_fields.setdefault("is_superuser", False)
-
-        return self._create_user(email, password, **extra_fields)
-
-    def get_active(self):
-        return self.get_queryset().filter(is_active=True).prefetch_related("member")
-
-    def create_superuser(self, email, password=None, **extra_fields):
-        extra_fields.setdefault("is_staff", True)
-        extra_fields.setdefault("is_superuser", True)
-        extra_fields.setdefault("is_active", True)
-
-        assert extra_fields.get("is_staff"), "Superuser must be assigned to is_staff=True"
-        assert extra_fields.get("is_superuser"), "Superuser must be assigned to is_superuser=True"
-        assert extra_fields.get("is_active"), "Superuser must be assigned to is_active=True"
-
-        return self._create_user(email, password, **extra_fields)
-
-    def _create_user(self, email, password, **extra_fields):
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.password = make_password(password)
-        user.save()
-        return user
