+++ apps/questions/mapping.py
@@ -5,11 +5,11 @@ from questions.models import QuestionSingleAnswer, InfoSlide, QuestionConnect, Q
 
 
 class TypeQuestionPoints(Enum):
-    INFO_SLIDE = 10
-    QUESTION_SINGLE_ANSWER = 25
-    QUESTION_WRITE = 40
-    QUESTION_CONNECT = 50
-    QUESTION_EXCLUDE = 60
+    INFO_SLIDE = 5
+    QUESTION_SINGLE_ANSWER = 5
+    QUESTION_WRITE = 5
+    QUESTION_CONNECT = 5
+    QUESTION_EXCLUDE = 5
 
 
 TaskObjs = Union[
